// src/types/enums.ts
export enum YandexCapability {
  On_Off = 'devices.capabilities.on_off',
  Color_setting = 'devices.capabilities.color_setting',
  Video_stream = 'devices.capabilities.video_stream',
  Mode = 'devices.capabilities.mode',
  Range = 'devices.capabilities.range',
  Toggle = 'devices.capabilities.toggle',
}
export enum YandexProperty {
  Float = 'devices.properties.float',
  Event = 'devices.properties.event',
}